+++ pypeit/par/pypeitpar.py
@@ -1094,7 +1094,7 @@ class CubePar(ParSet):
     see :ref:`pypeitpar`.
     """
 
-    def __init__(self, slit_spec=None, cube_spat_num=None, cube_wave_num=None,
+    def __init__(self, slit_spec=None, make_cube=None, cube_spat_num=None, cube_wave_num=None,
                  cube_wave_min=None, cube_wave_max=None):
 
         # Grab the parameter names and values from the function
@@ -1117,6 +1117,10 @@ class CubePar(ParSet):
         dtypes['slit_spec'] = [bool]
         descr['slit_spec'] = 'If the data use slits in one spatial direction, set this to True.' \
                              'If the data uses fibres for all spaxels, set this to False.'
+
+        defaults['make_cube'] = True
+        dtypes['make_cube'] = [bool]
+        descr['make_cube'] = 'Set this to False if you do not wish to generate a data cube.'
 
         defaults['cube_spat_num'] = None
         dtypes['cube_spat_num'] = [int, float]
@@ -1155,7 +1159,7 @@ class CubePar(ParSet):
         k = numpy.array([*cfg.keys()])
 
         # Basic keywords
-        parkeys = ['slit_spec', 'cube_spat_num', 'cube_wave_num', 'cube_wave_min', 'cube_wave_max']
+        parkeys = ['slit_spec', 'make_cube', 'cube_spat_num', 'cube_wave_num', 'cube_wave_min', 'cube_wave_max']
 
         badkeys = numpy.array([pk not in parkeys for pk in k])
         if numpy.any(badkeys):